Ensure Makefile_OSX is up-to-date on macOS
Add openjdk from brew, make testing quicker by checking before cloning and patching cbflib

+++ releng/prepare_source.sh
@@ -1,8 +1,12 @@
 # 
-git clone --depth=1 --branch=CBFlib-0.9.7-devel https://github.com/yayahjb/cbflib.git
+if [ ! -d cbflib ]; then
+  git clone --depth=1 --branch=CBFlib-0.9.7-devel https://github.com/yayahjb/cbflib.git
+fi
 cd cbflib
-ln -s ../releng .
-patch -p1 < releng/make.patch
-patch -p1 < releng/no-hdf5.patch
-patch -p1 < releng/pointer-leak.patch
+if [ ! -L releng ]; then
+  ln -s ../releng .
+  patch -p1 < releng/make.patch
+  patch -p1 < releng/no-hdf5.patch
+  patch -p1 < releng/pointer-leak.patch
+fi
